Renamed portentaBreakoutAnalog to Analog
