+++ src/Arduino_PortentaBreakoutCarrier.h
@@ -31,7 +31,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
 #include <mbed.h>
 #include <Wire.h>
 #include <SPI.h>
-#include "utility/Analog/portentaBreakoutAnalog.h"
+#include "utility/Analog/Analog.h"
 
 #define LAST_ARDUINO_PIN_NUMBER LEDB + 1
 typedef enum {
